refactor: create card folder in component and move workspaceCard
